backend/requirements.txt
fastapi
uvicorn[standard]
anthropic
python-dotenv
python-multipart
PyPDF2
python-docx
supabase
PyJWT
httpx
slowapi
python-jobspy
